Move proc gen into folders
