+++ projects/VoxelRenderer/src/procgen/generators/TextureHeightmapWorldGenerator.cpp
@@ -1,6 +1,6 @@
+#include <src/procgen/generators/TextureHeightmapWorldGenerator.h>
 #include <PerlinNoise/PerlinNoise.hpp>
 #include <imgui/imgui.h>
-#include <src/procgen/TextureHeightmapWorldGenerator.h>
 #include <src/world/VoxelWorldData.h>
 
 void TextureHeightmapWorldGenerator::generateData()
