+++ projects/VoxelRenderer/src/procgen/synthesizers/PointSynthesizer.h
